Revert "revert: roll source tree back to v1.4.0 to restore production"
This reverts commit e444a90b10ee7d334b95fb60484ab1cd0514fe16.
