+++ messages/en.json
@@ -62,7 +62,7 @@
     "colMin": "Min",
     "colMax": "Max",
     "colN": "n",
-    "bpClassificationTitle": "Blood pressure — ESC/ESH classification (2018)",
+    "bpClassificationTitle": "Blood pressure — ESH classification (2023)",
     "avgBp": "Average blood pressure",
     "classification": "Classification",
     "bmiTitle": "Body Mass Index (BMI)",
@@ -634,7 +634,7 @@
     "inTargetRange": "in target range",
     "bpTarget": "Target: {range}",
     "noDobSet": "Date of birth not set.",
-    "avg30dEsc": "30d average · ESC/ESH 2018",
+    "avg30dEsc": "30d average · ESH 2023",
     "whoClassification": "WHO Classification",
     "correlations": "Correlations",
     "weightVsBp": "Weight vs. Blood Pressure (Systolic)",
